+++ src/Program.cs
@@ -72,6 +72,7 @@ internal class Program
             new AzureMcp.Areas.ServiceBus.ServiceBusSetup(),
             new AzureMcp.Areas.Storage.StorageSetup(),
             new AzureMcp.Areas.BicepSchema.BicepSchemaSetup(),
+            new AzureMcp.Areas.VirtualDesktop.VirtualDesktopSetup(),
         ];
     }